fish user login

+++ src/NFComm/NFPluginModule/NFICoroutineModule.h
@@ -30,7 +30,7 @@ public:
 
     }
 
-    virtual int MakeCoroutine(const std::function<void()> &func) = 0;
+    virtual int64_t MakeCoroutine(const std::function<void()> &func) = 0;
 
     /// @brief 启动该协程任务, 执行Run方法
     /// @param is_immediately 是否立即执行
