+++ src/NFCommPlugin/NFKernelPlugin/NFCMessageModule.cpp
@@ -380,7 +380,7 @@ int NFCMessageModule::OnHandleReceiveNetPack(uint64_t connectionLink, uint64_t o
             if (callBack.mxAllMsgCallBackList.m_createCo)
             {
                 NET_RECEIVE_FUNCTOR &pFun = callBack.mxAllMsgCallBackList.m_pFunctor;
-                int coId = FindModule<NFICoroutineModule>()->MakeCoroutine(
+                int64_t coId = FindModule<NFICoroutineModule>()->MakeCoroutine(
                         [pFun, objectLinkId, packet]
                         {
                             //从消息层传过来的包中的数据，会在处理函数执行完后销毁掉，所以携程必须复制一份，以防万一yield后又用到。
@@ -415,7 +415,7 @@ int NFCMessageModule::OnHandleReceiveNetPack(uint64_t connectionLink, uint64_t o
                 int iRet = 0;
                 if (netFunctor.m_createCo)
                 {
-                    int coId = FindModule<NFICoroutineModule>()->MakeCoroutine(
+                    int64_t coId = FindModule<NFICoroutineModule>()->MakeCoroutine(
                             [objectLinkId, packet, pFun]
                             {
                                 //从消息层传过来的包中的数据，会在处理函数执行完后销毁掉，所以携程必须复制一份，以防万一yield后又用到。
@@ -474,7 +474,7 @@ int NFCMessageModule::OnHandleReceiveNetPack(uint64_t connectionLink, uint64_t o
                 int iRet = 0;
                 if (iterator2->second.m_createCo)
                 {
-                    int coId = FindModule<NFICoroutineModule>()->MakeCoroutine(
+                    int64_t coId = FindModule<NFICoroutineModule>()->MakeCoroutine(
                             [objectLinkId, packet, pFun]
                             {
                                 //从消息层传过来的包中的数据，会在处理函数执行完后销毁掉，所以携程必须复制一份，以防万一yield后又用到。
@@ -796,7 +796,7 @@ int NFCMessageModule::OnHandleRpcService(uint64_t connectionLink, uint64_t objec
                 if (netRpcService.m_createCo)
                 {
                     NFIRpcService* pRpcService = netRpcService.m_pRpcService;
-                    int coId = FindModule<NFICoroutineModule>()->MakeCoroutine(
+                    int64_t coId = FindModule<NFICoroutineModule>()->MakeCoroutine(
                             [this, pRpcService, objectLinkId, reqSvrPkg]()
                             {
                                 int iRet = pRpcService->run(objectLinkId, reqSvrPkg);
@@ -893,7 +893,7 @@ int NFCMessageModule::OnSocketNetEvent(eMsgType nEvent, uint64_t serverLinkId, u
             {
                 if (iter->second.m_createCo)
                 {
-                    int coId = FindModule<NFICoroutineModule>()->MakeCoroutine(
+                    int64_t coId = FindModule<NFICoroutineModule>()->MakeCoroutine(
                             [pFun, nEvent, objectLinkId]{
                                 pFun(nEvent, objectLinkId);
                             });
